Update hw5 directions
