--- hw5/_submissions/.gitignore
@@ -1,1 +1,0 @@
-.DS_Store
