+++ hw5/_submissions/drean-patrik/mumbling/README.md
@@ -1,0 +1,22 @@
+# Mumbling
+
+## Setup
+
+Copy this folder and paste it under `_submissions/lastname-firstname`.
+
+Now in the new directory, run `npm install` in order to install the required packages.
+
+The code is located in the `src` folder.
+
+Run `npm run test` to verify if tests are passing or not. Remember, you have to be in the directory for this command to work.
+
+## Instructions
+
+Complete the function `createMumble(str)`. The following examples show the mumble pattern.
+
+```
+createMumble("abcd")    # "A-Bb-Ccc-Dddd"
+createMumble("RqaEzty") # "R-Qq-Aaa-Eeee-Zzzzz-Tttttt-Yyyyyyy"
+createMumble("cwAt")    # "C-Ww-Aaa-Tttt"
+createMumble("")        # ""
+```
