+++ hw5/_submissions/drean-patrik/mumbling/src/mumbling.js
@@ -1,0 +1,4 @@
+export const createMumble = str => {
+  // TODO
+  return undefined
+}
