+++ hw5/_submissions/drean-patrik/mumbling/src/mumbling.test.js
@@ -1,0 +1,11 @@
+import { createMumble } from './mumbling'
+
+describe('When mumbling words', () => {
+  it('should return an empty string when given an empty string', () => {
+    const result = createMumble('')
+    const expected = []
+
+    expect(result).toEqual(expected)
+  })
+  // TODO: Add more tests
+})
