+++ hw5/_submissions/drean-patrik/scramblies/src/scramblies.js
@@ -1,0 +1,4 @@
+export const scramble = (str1, str2) => {
+  // TODO
+  return undefined
+}
