+++ hw5/_submissions/drean-patrik/scramblies/src/scramblies.test.js
@@ -1,0 +1,11 @@
+import { scramble } from './scramblies'
+
+describe('When validating scrambled words', () => {
+  it('should return true for words with the same exact letters', () => {
+    const result = scramble('owrdl', 'world')
+    const expected = true
+
+    expect(result).toEqual(expected)
+  })
+  // TODO: Add more tests
+})
