+++ hw5/mumbling/README.md
@@ -2,11 +2,13 @@
 
 ## Setup
 
-In this directory, run `npm install` in order to install the required packages.
+Copy this folder and paste it under `_submissions/lastname-firstname`.
+
+Now in the new directory, run `npm install` in order to install the required packages.
 
 The code is located in the `src` folder.
 
-Run `npm run test` to verify if tests are passing or not.
+Run `npm run test` to verify if tests are passing or not. Remember, you have to be in the directory for this command to work.
 
 ## Instructions
 
@@ -16,4 +18,5 @@ Complete the function `createMumble(str)`. The following examples show the mumbl
 createMumble("abcd")    # "A-Bb-Ccc-Dddd"
 createMumble("RqaEzty") # "R-Qq-Aaa-Eeee-Zzzzz-Tttttt-Yyyyyyy"
 createMumble("cwAt")    # "C-Ww-Aaa-Tttt"
+createMumble("")        # ""
 ```
