+++ hw5/mumbling/src/mumbling.js
@@ -1,4 +1,4 @@
 export const createMumble = str => {
   // TODO
-  return str
+  return undefined
 }
