+++ hw5/scramblies/README.md
@@ -2,11 +2,13 @@
 
 ## Setup
 
-In this directory, run `npm install` in order to install the required packages.
+Copy this folder and paste it under `_submissions/lastname-firstname`.
+
+Now in the new directory, run `npm install` in order to install the required packages.
 
 The code is located in the `src` folder.
 
-Run `npm run test` to verify if tests are passing or not.
+Run `npm run test` to verify if tests are passing or not. Remember, you have to be in the directory for this command to work.
 
 ## Instructions
