+++ hw5/scramblies/src/scramblies.js
@@ -1,4 +1,4 @@
 export const scramble = (str1, str2) => {
   // TODO
-  return false
+  return undefined
 }
